refactor(desktop): extract progress file actions and settings IPC helpers
Move the file-action concern out of DesktopProgressBridge into a focused
class, lift closure-independent settings IPC helpers into a sibling
module, and replace nested ternaries with early-return helpers.
